feat: - Menambahakan fitur CRUD data member dengan gambar profile
- Mengubah field pada data member (role dan divisi menggunakan enum)

- Memperbaiki Route koneksi database untuk sesi development dan production

+++ server/controllers/memberController.js
@@ -1,15 +1,34 @@
 import memberModel from "../models/memberModel.js";
 import attendanceRecordModel from "../models/attendanceRecordModel.js";
+import { v2 as cloudinary } from "cloudinary";
 
 //? Tambah Anggota baru ke master list
 export const addMember = async (req, res) => {
-  const { name, nim, divisi } = req.body;
-  if (!name) {
-    return res.json({ success: false, message: "Nama wajib diisi" });
+  const { name, role, nim, divisi } = req.body;
+  const image = req.file;
+
+  if (!image) {
+    return res.json({ success: false, message: "Gambar belum diupload" });
+  }
+
+  //! Path gambar yang akan disimpan di database
+  const imageUrl = req.file.path;
+  const public_id = req.file.filename;
+
+  if (!name || !role || !nim || !divisi) {
+    console.log(name, role, nim, divisi);
+    return res.json({ success: false, message: "Data tidak lengkap!" });
   }
 
   try {
-    const newMember = new memberModel({ name, nim, divisi });
+    const newMember = new memberModel({
+      public_id,
+      name,
+      role,
+      nim,
+      divisi,
+      imageUrl,
+    });
     await newMember.save();
     res.json({ success: true, message: "Anggota berhasil ditambahkan" });
   } catch (error) {
@@ -23,7 +42,7 @@ export const addMember = async (req, res) => {
 
 export const updateMember = async (req, res) => {
   const { id } = req.params;
-  const { name, nim, divisi } = req.body;
+  const { name, nim, divisi, role } = req.body;
 
   try {
     const member = await memberModel.findById(id);
@@ -35,7 +54,17 @@ export const updateMember = async (req, res) => {
       name,
       nim,
       divisi,
+      role,
     };
+
+    if (req.file) {
+      if (member.public_id) {
+        await cloudinary.uploader.destroy(member.public_id);
+      }
+
+      updateData.imageUrl = req.file.path;
+      updateData.public_id = req.file.filename;
+    }
 
     await memberModel.findByIdAndUpdate(id, updateData);
     res.json({ success: true, message: "Member berhasil diperbarui" });
@@ -48,7 +77,18 @@ export const updateMember = async (req, res) => {
 export const deleteMember = async (req, res) => {
   const { id } = req.params;
   try {
+    const member = await memberModel.findById(id);
+    if (!member)
+      return res.json({ success: false, message: "Data Memeber Tidak ada" });
+
+    //? Hapus data gambar jika ada
+    if (member.public_id) {
+      await cloudinary.uploader.destroy(member.public_id);
+    }
+
+    //! Hapus Data member
     await memberModel.findByIdAndDelete(id);
+
     //Todo Juga hapus semua record absensi terkait anggota ini
     await attendanceRecordModel.deleteMany({ memberId: id });
     res.json({ success: true, message: "Anggota berhasil dihapus" });
@@ -97,7 +137,7 @@ export const getAllMembers = async (req, res) => {
 
       //Todo hitung presentase kehadiran: (hadir / totalAcara) * 100
       const attendancePercentage =
-        totalRecords > 0 ? ((totalHadir / totalRecords) * 100).toFixed(0) : 0;
+        totalRecords > 0 ? ((totalHadir / totalRecords) * 100).toFixed(0) : 100;
 
       return {
         ...member,
